remove project/src
